Started work on a bash script for grabbing the assets off of a mounted mordor cd.
This is most of the way there, it checks to see if source files
exists, it makes sure the target directories exist, or sets them up if
they don't, and it echos the commands that would need to be executed in
order to do the setup.

Really, it just needs some verbosity flags, some dry run flags, and
maybe some nicer whitespacing before I try and push it up stream.