Read the total number of records in the file, printing out relevant parts as they're read in.
That I can read through all of the records, checking the alignment
pegs in the file suggest I'm reading in the right amount of data for
every element in the record, but some of the output doesn't make a lot
of sense.  In particular, I think the stat sets are signed (or maybe
sometimes signed).  It's unlikely that a cursed item adds 2^8 wisdom
to you, for example.

So, todo here

Use the spell cross references
Debug some of the input
Clean up